Refactor property and user DTOs; add new fields for property assignment and update event constants; enhance email types and templates for property assignment notifications.
