+++ apps/klubiq-dashboard/src/properties/dto/requests/property-manager.dto.ts
@@ -1,4 +1,9 @@
-export class PropertyManagerDto {
+export class PropertyManagerAssignmentDto {
 	uid: string;
 	isPropertyOwner: boolean;
+	propertyName: string;
+	name: string;
+	email: string;
+	orgName: string;
+	propertyAddress: string;
 }
